v2-dev/requirements.txt
discord.py
aiosqlite
python-dotenv
PyYAML
aiohttp
yoyo-migrations
psutil
asyncio